requirements.txt
pandas==2.2.2
requests==2.32.3
pyarrow==16.1.0
easydict==1.13
google-auth==2.31.0
google-cloud-bigquery==3.25.0
